Fix remaining API pitfalls documented in the integration guide
Errors that were 500 or misleading:
- Invalid course program JSON returns 400 "course program must be a JSON
  array" (json.Unmarshal accepts "null" into a slice, so a nil check is
  needed on top of the error).
- Translation problems return specific 400 messages instead of a generic
  "invalid request body"; the errors still wrap ErrInvalidInput.
- A student with a missing company returns 404 "company not found".
- A taken email returns 409 on user create, admin update and profile
  update.
- Subordinate lists of a missing parent return 404. The parent is looked up
  only when the list is empty, so the common case costs no extra query.
  Audit logs keep returning an empty list so history stays readable after
  e.g. a user is deleted.

Constraint names are matched as sets: databases created from the old
schema carry both the legacy constraints (fk_company, unique_email) and
the ones added by migrations, and PostgreSQL reports whichever it checks
first. Matching a single name passed unit tests but failed on a real DB.

Behavior:
- PATCH /courses/{id} without certificateTranslations leaves translations
  untouched; an explicit [] removes them. Omitting the field used to wipe
  every translation.
- Journal creation and session generation run in a transaction that rolls
  back when the last session falls after dateEnd, returning 400 "course
  program does not fit within journal dates". The scheduling SQL stays in
  one place; the check inspects its result. The handler requires the
  service (compile-time assertion) rather than falling back to an
  unchecked write through the querier.
- Adding an attendee to a closed journal returns 409, matching removal.
- Company NIPs are validated and stored as digits only, the format every
  existing row already uses. ValidateNIP now rejects non-digits, which it
  previously fed into the checksum.
- Dashboard expiry uses the certificate snapshot, like GET /certificates,
  and no longer drops students without a company. The count and the list
  now share the same conditions.

Response shapes (web/ updated in the same commit; deploy together):
certificate details expose studentFirstname and null for missing optional
values; PATCH /certificates returns full printVariants; company list
contactPerson is null instead of omitted; session hours and dashboard
registryNumber are numbers. The GUS lookup 404 message becomes "company
not found in GUS registry" so it no longer collides with the new
"company not found".

PESEL is deliberately left unvalidated: most stored values are other
identity document numbers.

Verification: 40 new tests; rejected write paths exercised on a running
API with checksums over nine tables unchanged; PATCH on a real course
confirmed the translation row, including updated_at, was untouched; the
dashboard queries matched the old count and every listed expiry date; the
spec re-validated against 61 live calls; server-rendered pages checked.

+++ api/internal/dashboard/dto.go
@@ -21,12 +21,12 @@ type ExpiringSummaryDTO struct {
 }
 
 type ExpiringCertificateDTO struct {
-	CertificateID  int64   `json:"certificateId"`
-	ExpiryDate     string  `json:"expiryDate"`
-	StudentName    string  `json:"studentName"`
-	CompanyName    string  `json:"companyName"`
-	CourseName     string  `json:"courseName"`
-	CourseSymbol   string  `json:"courseSymbol"`
-	RegistryYear   int64   `json:"registryYear"`
-	RegistryNumber float64 `json:"registryNumber"`
+	CertificateID  int64  `json:"certificateId"`
+	ExpiryDate     string `json:"expiryDate"`
+	StudentName    string `json:"studentName"`
+	CompanyName    string `json:"companyName"`
+	CourseName     string `json:"courseName"`
+	CourseSymbol   string `json:"courseSymbol"`
+	RegistryYear   int64  `json:"registryYear"`
+	RegistryNumber int64  `json:"registryNumber"`
 }
